Bike Rental Prediction
The objective of this Case is to Predication of bike rental count on daily based on the environmental and seasonal settings. Predicting demand of ride sharing systems has become a common problem in the modern world, with companies such as Uber and Ola emerging in the car-sharing service. Similarly, bike-sharing services have gained considerable traction in the past decade. Heavy street traffic in busy cities and a desire for an environmentally friendly form of transportation make biking an attractive alternative to traveling by car.